Passes command decoration tests within hero section, adds failing tests for header section, refactors tests
